requirements.txt
Django==4.1.2
django-filter==22.1
djangorestframework==3.13.1
sqlparse==0.4.3
django-cors-headers==3.13.0
django-environ==0.9.0
psycopg2
environ
gspread